requirements.txt
asgiref==3.5.2
Django==4.1.4
django-environ==0.9.0
django-extensions==3.2.1
django-model-utils==4.3.1
flake8==6.0.0
gunicorn==20.1.0
isort==5.11.4
psycopg2-binary==2.9.5
sqlparse==0.4.3
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1

